+++ controllers/SiteController.php
@@ -7,7 +7,7 @@ use Yii;
 use yii\filters\AccessControl;
 use yii\web\Response;
 use yii\filters\VerbFilter;
-use app\models\LoginForm;
+use app\models\forms\LoginForm;
 
 class SiteController extends MainController
 {
